chrome/content/mochitest/tests/test_filterStorage.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Filter storage tests</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>        

  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">

  </div>
  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    //TODO: increaseHitCount, resetHitCounts

    prepareFilterComponents();

    function compareSubscriptionObservers(test, list)
    {
      let result = getGlobalForObject(FilterStorage).subscriptionObservers.map(function(observer) {return observer.name}).join("\n");
      let expected = list.map(function(observer) {return observer.name}).join("\n");
      is(result, expected, test);
    }

    function compareFilterObservers(test, list)
    {
      let result = getGlobalForObject(FilterStorage).filterObservers.map(function(observer) {return observer.name}).join("\n");
      let expected = list.map(function(observer) {return observer.name}).join("\n");
      is(result, expected, test);
    }

    function compareSubscriptionList(test, list)
    {
      let result = FilterStorage.subscriptions.map(function(subscription) {return subscription.url}).join("\n");
      let expected = list.map(function(subscription) {return subscription.url}).join("\n");
      is(result, expected, test);
    }

    function compareFiltersList(test, list)
    {
      let result = FilterStorage.subscriptions.map(function(subscription) {return subscription.filters.map(function(filter) {return filter.text}).join("\n") + "\n"}).join("\n");
      let expected = list.map(function(filters) {return filters.join("\n") + "\n"}).join("\n");
      is(result, expected, test);
    }

    function compareFilterSubscriptions(test, filter, list)
    {
      let result = filter.subscriptions.map(function(subscription) {return subscription.url}).join("\n");
      let expected = list.map(function(subscription) {return subscription.url}).join("\n");
      is(result, expected, test);
    }

    testSubscriptions();
    testFilterSubscriptionRelationship();

    FilterStorage.addSubscription(Subscription.fromURL("~fl~"));
    FilterStorage.addSubscription(Subscription.fromURL("~wl~"));
    FilterStorage.addSubscription(Subscription.fromURL("~il~"));
    FilterStorage.addSubscription(Subscription.fromURL("~eh~"));

    testFilters();
    testDisabledSubscriptions();

    function testSubscriptions()
    {
      let observed1 = "";
      let observed2 = "";
      let observed3 = "";
      let observer1 = function subscription_observer1(action, subscriptions) { observed1 += action + " " + subscriptions.map(function(subscription) {return subscription.url;}).join(" ") + "\n"; }
      let observer2 = function subscription_observer2(action, subscriptions) { observed2 += action + " " + subscriptions.map(function(subscription) {return subscription.url;}).join(" ") + "\n"; }
      let observer3 = function subscription_observer3(action, subscriptions) { observed3 += action + " " + subscriptions.map(function(subscription) {return subscription.url;}).join(" ") + "\n"; }

      let subscription1 = Subscription.fromURL("http://test1/");
      let subscription2 = Subscription.fromURL("http://test2/");
      let subscription3 = Subscription.fromURL("http://test3/");

      compareSubscriptionList("No subscriptions", []);
      compareSubscriptionObservers("No observers", []);

      FilterStorage.addSubscription(subscription1);
      compareSubscriptionList("add(test1)", [subscription1]);

      FilterStorage.addSubscriptionObserver(observer1);
      compareSubscriptionObservers("addSubscriptionObserver(observer1)", [observer1]);

      FilterStorage.addSubscriptionObserver(observer1);
      compareSubscriptionObservers("addSubscriptionObserver(observer1) again", [observer1]);

      FilterStorage.addSubscription(subscription1);
      compareSubscriptionList("add(test1) again", [subscription1]);

      FilterStorage.removeSubscription(subscription2);
      compareSubscriptionList("remove(test2)", [subscription1]);

      FilterStorage.addSubscriptionObserver(observer2);
      compareSubscriptionObservers("addSubscriptionObserver(observer2)", [observer1, observer2]);

      FilterStorage.removeSubscriptionObserver(observer1);
      compareSubscriptionObservers("removeSubscriptionObserver(observer1)", [observer2]);

      FilterStorage.addSubscription(subscription2, true);
      compareSubscriptionList("add(test2) silent", [subscription1, subscription2]);

      FilterStorage.addSubscription(subscription3);
      compareSubscriptionList("add(test3)", [subscription1, subscription2, subscription3]);

      FilterStorage.addSubscriptionObserver(observer3);
      compareSubscriptionObservers("addSubscriptionObserver(observer3)", [observer2, observer3]);

      FilterStorage.removeSubscription(subscription1, true);
      compareSubscriptionList("remove(test1) silent", [subscription2, subscription3]);

      FilterStorage.addSubscriptionObserver(observer1);
      compareSubscriptionObservers("addSubscriptionObserver(observer1)", [observer2, observer3, observer1]);

      FilterStorage.removeSubscriptionObserver(observer3);
      compareSubscriptionObservers("removeSubscriptionObserver(observer3)", [observer2, observer1]);

      FilterStorage.removeSubscription(subscription1);
      compareSubscriptionList("remove(test1) again", [subscription2, subscription3]);

      FilterStorage.addSubscription(subscription1);
      compareSubscriptionList("add(test1)", [subscription2, subscription3, subscription1]);

      FilterStorage.removeSubscription(subscription3);
      compareSubscriptionList("remove(test3)", [subscription2, subscription1]);

      FilterStorage.removeSubscription(subscription3);
      compareSubscriptionList("remove(test3) again", [subscription2, subscription1]);

      is(observed1, "add http://test1/\nremove http://test3/\n", "observed1");
      is(observed2, "add http://test3/\nadd http://test1/\nremove http://test3/\n", "observed2");
      is(observed3, "", "observed3");

      FilterStorage.removeSubscription(subscription1);
      FilterStorage.removeSubscription(subscription2);
      FilterStorage.removeSubscription(subscription3);
      FilterStorage.removeSubscriptionObserver(observer1);
      FilterStorage.removeSubscriptionObserver(observer2);
      FilterStorage.removeSubscriptionObserver(observer3);
    }

    function testFilters()
    {
      let observed1 = "";
      let observed2 = "";
      let observed3 = "";
      let observer1 = function filter_observer1(action, filters) { observed1 += action + " " + filters.map(function(filter) {return filter.text;}).join(" ") + "\n"; }
      let observer2 = function filter_observer2(action, filters) { observed2 += action + " " + filters.map(function(filter) {return filter.text;}).join(" ") + "\n"; }
      let observer3 = function filter_observer3(action, filters) { observed3 += action + " " + filters.map(function(filter) {return filter.text;}).join(" ") + "\n"; }

      compareFiltersList("No filters", [[],[],[],[]]);
      compareFilterObservers("No observers", []);

      FilterStorage.addFilter(Filter.fromText("foo"));
      compareFiltersList("add(foo)", [["foo"],[],[],[]]);

      FilterStorage.addFilterObserver(observer1);
      compareFilterObservers("addFilterObserver(observer1)", [observer1]);

      FilterStorage.addFilterObserver(observer1);
      compareFilterObservers("addFilterObserver(observer1) again", [observer1]);

      FilterStorage.addFilter(Filter.fromText("bar"));
      compareFiltersList("add(bar)", [["foo", "bar"],[],[],[]]);

      FilterStorage.removeFilter(Filter.fromText("foobar"));
      compareFiltersList("remove(foobar)", [["foo", "bar"],[],[],[]]);

      FilterStorage.addFilterObserver(observer2);
      compareFilterObservers("addFilterObserver(observer2)", [observer1, observer2]);

      FilterStorage.removeFilterObserver(observer1);
      compareFilterObservers("removeFilterObserver(observer1)", [observer2]);

      FilterStorage.addFilter(Filter.fromText("@@ddd"));
      compareFiltersList("add(@@ddd)", [["foo", "bar"],["@@ddd"],[],[]]);

      FilterStorage.addFilter(Filter.fromText("!foobar"));
      compareFiltersList("add(!foobar)", [["foo", "bar", "!foobar"],["@@ddd"],[],[]]);

      FilterStorage.addFilterObserver(observer3);
      compareFilterObservers("addFilterObserver(observer3)", [observer2, observer3]);

      FilterStorage.removeFilter(Filter.fromText("bar"));
      compareFiltersList("remove(bar)", [["foo", "!foobar"],["@@ddd"],[],[]]);

      FilterStorage.removeFilter(Filter.fromText("@@ddd"));
      compareFiltersList("remove(@@ddd)", [["foo", "!foobar"],[],[],[]]);

      FilterStorage.addFilterObserver(observer1);
      compareFilterObservers("addFilterObserver(observer1)", [observer2, observer3, observer1]);

      FilterStorage.removeFilterObserver(observer3);
      compareFilterObservers("removeFilterObserver(observer3)", [observer2, observer1]);

      FilterStorage.addFilter(Filter.fromText("foo#bar"));
      compareFiltersList("add(foo#bar)", [["foo", "!foobar"],[],[],["foo#bar"]]);

      FilterStorage.addFilter(Filter.fromText("/??/"));
      compareFiltersList("add(/??/)", [["foo", "!foobar"],[],["/??/"],["foo#bar"]]);

      FilterStorage.addFilter(Filter.fromText("@@asdf"));
      compareFiltersList("add(@asdf/)", [["foo", "!foobar"],["@@asdf"],["/??/"],["foo#bar"]]);

      is(observed1, "add bar\nadd foo#bar\nadd /??/\nadd @@asdf\n", "observed1");
      is(observed2, "add @@ddd\nadd !foobar\nremove bar\nremove @@ddd\nadd foo#bar\nadd /??/\nadd @@asdf\n", "observed2");
      is(observed3, "remove bar\nremove @@ddd\n", "observed3");

      FilterStorage.removeFilterObserver(observer1);
      FilterStorage.removeFilterObserver(observer2);
      FilterStorage.removeFilterObserver(observer3);

      // Tests for adding filters before a particular filter

      FilterStorage.addFilter(Filter.fromText("test1"), Filter.fromText("!foobar"));
      compareFiltersList("add(test1) before !foobar", [["foo", "test1", "!foobar"],["@@asdf"],["/??/"],["foo#bar"]]);

      FilterStorage.addFilter(Filter.fromText("test2"), Filter.fromText("foo"));
      compareFiltersList("add(test2) before foo", [["test2", "foo", "test1", "!foobar"],["@@asdf"],["/??/"],["foo#bar"]]);

      FilterStorage.addFilter(Filter.fromText("test3"), Filter.fromText("@@asdf"));
      compareFiltersList("add(test3) before @@asdf", [["test2", "foo", "test1", "!foobar", "test3"],["@@asdf"],["/??/"],["foo#bar"]]);

      FilterStorage.addFilter(Filter.fromText("@@test4"), Filter.fromText("@@asdf"));
      compareFiltersList("add(@@test5) before @@asdf", [["test2", "foo", "test1", "!foobar", "test3"],["@@test4", "@@asdf"],["/??/"],["foo#bar"]]);
    }

    function testFilterSubscriptionRelationship()
    {
      let filter1 = Filter.fromText("filter1");
      let filter2 = Filter.fromText("filter2");
      let filter3 = Filter.fromText("filter3");

      let subscription1 = Subscription.fromURL("http://test1/");
      subscription1.filters = [filter1, filter2];

      let subscription2 = Subscription.fromURL("http://test2/");
      subscription2.filters = [filter2, filter3];

      let subscription3 = Subscription.fromURL("http://test3/");
      subscription3.filters = [filter1, filter2, filter3];

      compareFilterSubscriptions("Initial filter1 subscriptions", filter1, []);
      compareFilterSubscriptions("Initial filter2 subscriptions", filter2, []);
      compareFilterSubscriptions("Initial filter3 subscriptions", filter3, []);

      FilterStorage.addSubscription(subscription1);

      compareFilterSubscriptions("filter1 subscriptions after adding http://test1/", filter1, [subscription1]);
      compareFilterSubscriptions("filter2 subscriptions after adding http://test1/", filter2, [subscription1]);
      compareFilterSubscriptions("filter3 subscriptions after adding http://test1/", filter3, []);

      FilterStorage.addSubscription(subscription2);

      compareFilterSubscriptions("filter1 subscriptions after adding http://test2/", filter1, [subscription1]);
      compareFilterSubscriptions("filter2 subscriptions after adding http://test2/", filter2, [subscription1, subscription2]);
      compareFilterSubscriptions("filter3 subscriptions after adding http://test2/", filter3, [subscription2]);

      FilterStorage.removeSubscription(subscription1);

      compareFilterSubscriptions("filter1 subscriptions after removing http://test1/", filter1, []);
      compareFilterSubscriptions("filter2 subscriptions after removing http://test1/", filter2, [subscription2]);
      compareFilterSubscriptions("filter3 subscriptions after removing http://test1/", filter3, [subscription2]);

      FilterStorage.updateSubscriptionFilters(subscription3, [filter3]);

      compareFilterSubscriptions("filter1 subscriptions after updating http://test3/ filters", filter1, []);
      compareFilterSubscriptions("filter2 subscriptions after updating http://test3/ filters", filter2, [subscription2]);
      compareFilterSubscriptions("filter3 subscriptions after updating http://test3/ filters", filter3, [subscription2]);

      FilterStorage.addSubscription(subscription3);

      compareFilterSubscriptions("filter1 subscriptions after adding http://test3/", filter1, []);
      compareFilterSubscriptions("filter2 subscriptions after adding http://test3/", filter2, [subscription2]);
      compareFilterSubscriptions("filter3 subscriptions after adding http://test3/", filter3, [subscription2, subscription3]);

      FilterStorage.updateSubscriptionFilters(subscription3, [filter1, filter2]);

      compareFilterSubscriptions("filter1 subscriptions after updating http://test3/ filters", filter1, [subscription3]);
      compareFilterSubscriptions("filter2 subscriptions after updating http://test3/ filters", filter2, [subscription2, subscription3]);
      compareFilterSubscriptions("filter3 subscriptions after updating http://test3/ filters", filter3, [subscription2]);

      FilterStorage.removeSubscription(subscription3);

      compareFilterSubscriptions("filter1 subscriptions after removing http://test3/", filter1, []);
      compareFilterSubscriptions("filter2 subscriptions after removing http://test3/", filter2, [subscription2]);
      compareFilterSubscriptions("filter3 subscriptions after removing http://test3/", filter3, [subscription2]);

      FilterStorage.removeSubscription(subscription1);
      FilterStorage.removeSubscription(subscription2);
      FilterStorage.removeSubscription(subscription3);
    }

    function testDisabledSubscriptions()
    {
      let subscription = Subscription.fromURL("~fl~");
      FilterStorage.updateSubscriptionFilters(subscription, []);
      subscription.disabled = true;

      let filter1 = Filter.fromText("filter1");
      let filter2 = Filter.fromText("filter2");

      FilterStorage.addFilter(filter1);
      is(subscription.disabled, true, "Subscription still disabled after adding one filter");
      FilterStorage.addFilter(filter2);
      is(subscription.disabled, true, "Subscription still disabled after adding two filters");

      FilterStorage.removeFilter(filter1);
      is(subscription.disabled, true, "Subscription still disabled after removing one filter");
      FilterStorage.removeFilter(filter2);
      is(subscription.disabled, false, "Subscription no longer disabled after removing all filters");

      subscription.disabled = true;
      FilterStorage.updateSubscriptionFilters(subscription, [filter1]);
      is(subscription.disabled, true, "Subscription still disabled after replacing filters by one filter");
      FilterStorage.updateSubscriptionFilters(subscription, [filter1, filter2]);
      is(subscription.disabled, true, "Subscription still disabled after replacing filters by two filters");
      FilterStorage.updateSubscriptionFilters(subscription, []);
      is(subscription.disabled, false, "Subscription no longer disabled after replacing filters by an empty list");
    }
  </script>
  </pre>
</body>
</html>
